fuckin broken still WHY WHY WHY gd it
problems:
- casting is not doing for some reason
  (why it not do!!???)

updates:
this branch is for the real version of recipe game.
- created a bunch of neat assets
- set up the basic structure for how imma have multiple recipes
- redid the scoring system so it's not a salt castle. now ingredients are sanctioned correctly. (sanctioned? correct word?)
- stopped using mfkin spritebuilder (seriously, why even) for the ccb files that denoted CCNodes. bro i can make my own ccnode kthx.

disclaimer:
this commit message is abnormally long, even for me. it's because it's friday and i pregamed this. yep. i pregamed debugging session. why doesn't everyone do that.
also i made a lot of changes that weren't broken up into smaller commits because shit was broken and shit still be broken. enjoy this shitshow.